requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
blinker==1.9.0
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
ecdsa==0.19.1
fastapi==0.135.1
Flask==3.1.3
fonttools==4.62.1
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pillow==12.1.1
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-jose==3.5.0
rsa==4.9.1
seaborn==0.13.2
six==1.17.0
starlette==0.52.1
tabulate==0.10.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
Werkzeug==3.1.6
